🚚 move tracekit tests in the corresponding module spec
* `handler.spec` contained test cases that subscribed and threw uncaught
  error to see if they were correctly reported. That's the `report`
  module logic, so it has moved there.

* `parser.spec` contained test cases using `computeStackTrace`, so they
  has been moved to the `computeStackTrace` module.
